requirements.txt
pyside6
pyobjc-framework-Quartz
opencv-python
pynput
pytest
numpy
